requirements.txt
requests
pywin32
Pillow
pyautogui